confirm, layout nonmember 수정

+++ src/components/CalendarMonthWithoutTime.js
@@ -52,7 +52,19 @@ const CalendarMonthWithoutTime = (props) => {
 	const [removingDate, setRemovingDate] = useState(new Set());
 	const [adding, setAdding] = useState(true);
 
-	const updateAndDelete = async (adding, selectingDate) => {
+	let confirming = props.confirming;
+	let confirmSelected = props.confirmSelected;
+	let setConfirmSelected = props.setConfirmSelected;
+	let confirmed = props.confirmed;
+
+	let showConfirmModal = props.showConfirmModal;
+	let setShowConfirmModal = props.setShowConfirmModal;
+	let setModalPosition = props.setModalPosition;
+
+	const [confirmAdding, setConfirmAdding] = useState(new Set()); // 확정 선택 중 목록
+	const [confirmRemoving, setConfirmRemoving] = useState(new Set()); // 확정 선택 중 목록
+
+	const updateAndDelete = async (adding, selectingDate, selectDate) => {
 		const result = [];
 		if (adding) {
 			selectingDate
@@ -72,47 +84,49 @@ const CalendarMonthWithoutTime = (props) => {
 
 		console.log(result);
 		// 업데이트 로직 넣기
-		try {
-			if (adding) {
-				const response = await axios.post(
-					`${process.env.REACT_APP_API_URL}/promise/time`,
-					{
-						promiseId: promiseId.split('_')[0],
-						dateAvailable: result,
-					},
-					!accessToken && { headers: { Authorization: `@${nonmemberId}` } },
-				);
-				console.log(response.data);
-			} else {
-				if (accessToken) {
-					const response = await axios.delete(
-						`${process.env.REACT_APP_API_URL}/promise/deletetime`,
+		if (!confirming) {
+			try {
+				if (adding) {
+					const response = await axios.post(
+						`${process.env.REACT_APP_API_URL}/promise/time`,
 						{
-							data: {
-								promiseId: promiseId.split('_')[0],
-								dateToDelete: result,
-							},
+							promiseId: promiseId.split('_')[0],
+							dateAvailable: result,
 						},
+						!accessToken && { headers: { Authorization: `@${nonmemberId}` } },
 					);
 					console.log(response.data);
 				} else {
-					const response = await axios.delete(
-						`${process.env.REACT_APP_API_URL}/promise/deletetime`,
-						{
-							data: {
-								promiseId: promiseId.split('_')[0],
-								dateToDelete: result,
+					if (accessToken) {
+						const response = await axios.delete(
+							`${process.env.REACT_APP_API_URL}/promise/deletetime`,
+							{
+								data: {
+									promiseId: promiseId.split('_')[0],
+									dateToDelete: result,
+								},
 							},
-							headers: { Authorization: `@${nonmemberId}` },
-						},
-					);
-					console.log(response.data);
+						);
+						console.log(response.data);
+					} else {
+						const response = await axios.delete(
+							`${process.env.REACT_APP_API_URL}/promise/deletetime`,
+							{
+								data: {
+									promiseId: promiseId.split('_')[0],
+									dateToDelete: result,
+								},
+								headers: { Authorization: `@${nonmemberId}` },
+							},
+						);
+						console.log(response.data);
+					}
 				}
+				setReset(!reset);
+			} catch (error) {
+				const errorResponse = error.response;
+				console.log(errorResponse.data.statusCode);
 			}
-			setReset(!reset);
-		} catch (error) {
-			const errorResponse = error.response;
-			console.log(errorResponse.data.statusCode);
 		}
 		return result;
 	};
@@ -137,12 +151,27 @@ const CalendarMonthWithoutTime = (props) => {
 		if (!isDragging && (nonmemberId !== -1 || accessToken)) getMyParti();
 	}, [setReset, nonmemberId]);
 
-	const handleMouseDown = (date) => {
+	const handleMouseDown = (date, event) => {
 		if (editing) {
 			setDragStart(date);
 			setIsDragging(true);
 			if (selectDate.has(format(date, 'yyyy-MM-dd'))) setAdding(false);
 			else setAdding(true);
+		} else if (confirming) {
+			setDragStart(date);
+			setIsDragging(true);
+			if (confirmSelected.has(format(date, 'yyyy-MM-dd'))) setAdding(false);
+			else setAdding(true);
+		} else {
+			if (confirmed.has(format(date, 'yyyy-MM-dd'))) {
+				setTimeout(() => {
+					setShowConfirmModal('click');
+					console.log(event.clientX, event.clientY);
+					setModalPosition([event.clientX, event.clientY]);
+				}, 50);
+			} else {
+				// 참가자 고정?
+			}
 		}
 	};
 
@@ -157,7 +186,7 @@ const CalendarMonthWithoutTime = (props) => {
 					// selectingDate.difference(selectDate).forEach((dateStr) => {
 					// 	console.log(DaysOfWeek[getDay(new Date(dateStr))]);
 					// });
-					updateAndDelete(true, selectingDate);
+					updateAndDelete(true, selectingDate, selectDate);
 					setSelectDate(
 						(prevState) => new Set([...prevState, ...selectingDate]),
 					);
@@ -166,7 +195,7 @@ const CalendarMonthWithoutTime = (props) => {
 					// removingDate.intersection(selectDate).forEach((dateStr) => {
 					// 	console.log(DaysOfWeek[getDay(new Date(dateStr))]);
 					// });
-					updateAndDelete(false, removingDate);
+					updateAndDelete(false, removingDate, selectDate);
 
 					setSelectDate((prevState) => {
 						return new Set(
@@ -177,10 +206,32 @@ const CalendarMonthWithoutTime = (props) => {
 			}
 			setSelectingDate(new Set());
 			setRemovingDate(new Set());
+		} else if (confirming) {
+			setDragEnd(date);
+			setIsDragging(false);
+			if (isSameDay(dragStart, date)) onDateClick(date);
+			else {
+				if (adding) {
+					updateAndDelete(true, confirmAdding, confirmSelected);
+					setConfirmSelected(
+						(prevState) => new Set([...prevState, ...confirmAdding]),
+					);
+				} else {
+					updateAndDelete(false, confirmRemoving, confirmSelected);
+					setConfirmSelected((prevState) => {
+						return new Set(
+							[...prevState].filter((element) => !confirmRemoving.has(element)),
+						);
+					});
+				}
+			}
+			setConfirmAdding(new Set());
+			setConfirmRemoving(new Set());
 		}
+		console.log('confirmSelected', confirmSelected);
 	};
 
-	const handleMouseMove = (date) => {
+	const handleMouseMove = (date, event) => {
 		if (editing) {
 			if (isDragging && !isSameDay(dragEnd, date)) {
 				setDragEnd(date);
@@ -229,6 +280,66 @@ const CalendarMonthWithoutTime = (props) => {
 			}
 		} else {
 			getCanParti(date);
+			if (confirming) {
+				if (isDragging && !isSameDay(dragEnd, date)) {
+					setDragEnd(date);
+					setConfirmAdding(new Set());
+					setConfirmRemoving(new Set());
+
+					let start = new Date();
+					let end = new Date();
+					if (getDay(dragStart) <= getDay(date)) {
+						if (isBefore(dragStart, date)) {
+							// 오른쪽아래
+							start = dragStart;
+							end = date;
+						} else {
+							// 오른쪽위
+							start = addDays(startOfWeek(date), getDay(dragStart));
+							end = addDays(startOfWeek(dragStart), getDay(date));
+						}
+					} else {
+						if (isBefore(date, dragStart)) {
+							// 왼쪽위
+							start = date;
+							end = dragStart;
+						} else {
+							// 왼쪽아래
+							start = addDays(startOfWeek(dragStart), getDay(date));
+							end = addDays(startOfWeek(date), getDay(dragStart));
+						}
+					}
+
+					let A = start;
+					while (isBefore(A, addDays(end, 1))) {
+						let day = format(A, 'yyyy-MM-dd');
+						// console.log(day);
+						if (adding)
+							setConfirmAdding((prevState) => new Set([...prevState, day]));
+						// console.log(selectingDate)
+						else
+							setConfirmRemoving((prevState) => new Set([...prevState, day]));
+
+						if (getDay(A) == getDay(end)) {
+							A = addDays(A, 7 - getDay(end) + getDay(start));
+							continue;
+						}
+						A = addDays(A, 1);
+					}
+				}
+			} else {
+				if (confirmed.has(format(date, 'yyyy-MM-dd'))) {
+					if (showConfirmModal == 'no') {
+						setShowConfirmModal('hover');
+						console.log(event.clientX, event.clientY);
+						setModalPosition([event.clientX, event.clientY]);
+					}
+				} else {
+					if (showConfirmModal == 'hover') {
+						setShowConfirmModal('no');
+					}
+				}
+			}
 		}
 	};
 
@@ -288,6 +399,28 @@ const CalendarMonthWithoutTime = (props) => {
 									  !removingDate.has(format(day, 'yyyy-MM-dd'))
 										? `${styles.col} ${styles.day} ${styles.selected}`
 										: `${styles.col} ${styles.day} ${styles.valid}`
+									: confirming
+									? (confirmSelected.has(format(day, 'yyyy-MM-dd')) ||
+											confirmAdding.has(format(day, 'yyyy-MM-dd'))) &&
+									  !confirmRemoving.has(format(day, 'yyyy-MM-dd'))
+										? `${styles.col} ${styles.day} ${styles.confirmSelected}`
+										: color === 'valid'
+										? `${styles.col} ${styles.day} ${styles.valid}`
+										: color === 'E0CEFF'
+										? `${styles.col} ${styles.day} ${styles.Lv1}`
+										: color === 'BEA1FE'
+										? `${styles.col} ${styles.day} ${styles.Lv2}`
+										: color === 'A988F0'
+										? `${styles.col} ${styles.day} ${styles.Lv3}`
+										: color === '8D63E8'
+										? `${styles.col} ${styles.day} ${styles.Lv4}`
+										: color === '6330DE'
+										? `${styles.col} ${styles.day} ${styles.Lv5}`
+										: color === '4B1CBC'
+										? `${styles.col} ${styles.day} ${styles.Lv6}`
+										: `${styles.col} ${styles.day} ${styles.Lv7}`
+									: confirmed.has(format(day, 'yyyy-MM-dd'))
+									? `${styles.col} ${styles.day} ${styles.confirmSelected}`
 									: color === 'valid'
 									? `${styles.col} ${styles.day} ${styles.valid}`
 									: color === 'E0CEFF'
@@ -307,9 +440,9 @@ const CalendarMonthWithoutTime = (props) => {
 							// id={`${styles.selected}`}
 							style={getStyles()}
 							key={day}
-							onMouseDown={() => handleMouseDown(cloneDay)}
+							onMouseDown={(e) => handleMouseDown(cloneDay, e)}
 							onMouseUp={() => handleMouseUp(cloneDay)}
-							onMouseMove={() => handleMouseMove(cloneDay)}
+							onMouseMove={(e) => handleMouseMove(cloneDay, e)}
 						>
 							{!options.has(format(day, 'yyyy-MM-dd')) && (
 								<svg
@@ -441,18 +574,45 @@ const CalendarMonthWithoutTime = (props) => {
 		setSelectWeek(addMonths(selectWeek, 1));
 	};
 	const onDateClick = (day) => {
-		if (selectDate.has(format(day, 'yyyy-MM-dd'))) {
-			updateAndDelete(false, new Set([format(day, 'yyyy-MM-dd')]));
-			setSelectDate((prevState) => {
-				prevState.delete(format(day, 'yyyy-MM-dd'));
-				return new Set([...prevState]);
-			});
+		if (confirming) {
+			if (confirmSelected.has(format(day, 'yyyy-MM-dd'))) {
+				updateAndDelete(
+					false,
+					new Set([format(day, 'yyyy-MM-dd')]),
+					confirmSelected,
+				);
+				setConfirmSelected((prevState) => {
+					prevState.delete(format(day, 'yyyy-MM-dd'));
+					return new Set([...prevState]);
+				});
+			} else {
+				updateAndDelete(
+					true,
+					new Set([format(day, 'yyyy-MM-dd')]),
+					confirmSelected,
+				);
+				setConfirmSelected(
+					(prevState) => new Set([...prevState, format(day, 'yyyy-MM-dd')]),
+				);
+			}
 		} else {
-			updateAndDelete(true, new Set([format(day, 'yyyy-MM-dd')]));
-			// console.log(new Set([format(day, 'yyyy-MM-dd')]).difference(selectDate));
-			setSelectDate(
-				(prevState) => new Set([...prevState, format(day, 'yyyy-MM-dd')]),
-			);
+			if (selectDate.has(format(day, 'yyyy-MM-dd'))) {
+				updateAndDelete(
+					false,
+					new Set([format(day, 'yyyy-MM-dd')]),
+					selectDate,
+				);
+				setSelectDate((prevState) => {
+					prevState.delete(format(day, 'yyyy-MM-dd'));
+					return new Set([...prevState]);
+				});
+			} else {
+				updateAndDelete(true, new Set([format(day, 'yyyy-MM-dd')]), selectDate);
+				// console.log(new Set([format(day, 'yyyy-MM-dd')]).difference(selectDate));
+				setSelectDate(
+					(prevState) => new Set([...prevState, format(day, 'yyyy-MM-dd')]),
+				);
+			}
 		}
 	};
